Use lossy UTF-8 decoding for strings
This makes the decoding a little more tolerant to UTF-8 encoding
errors, which might not be the user's or contract author's fault.

One example is the string parameter in the `Apply` event on
https://etherscan.io/tx/0x13d400e7d06fe4b88d93a28d6974654604c3eb24b726f0671ee793add8b6c459#eventlog

It is not valid UTF-8, so something must've gone wrong on either the
user's side or in the contract. Rather than failing to decode this event
data, lossy UTF-8 allows us to decode it as "不�" and move on.